Local runs: tag executionSource + write SharedTestMemory
Two changes to the local action-script builders, extracted into a
config-free `action-script-builders.ts` module so the build logic is
unit-testable in isolation:

- `sharedTestMemoryId: projectId` (was `""`) — a non-empty id lets the
  studio persist run summaries + feedback to the project's
  SharedTestMemory on local runs; the memory writer skips on an empty id.

- `actionParams.executionSource: "local"` — tags the run so the studio
  can skip its own cloud ActionScript/TestScript write for MCP-local
  runs, leaving the /local-run/upload path as the single writer (the
  studio-side gate ships in muggle-ai-teaching-service). This removes the
  duplicate Firestore ActionScript doc per local run.
